### 2020-10-05 #2719 @mbobra
@Kevin-Mattheus-Moerman @arfon I am happy to edit! I think I can be an objective editor, but I do want to mention that I've used PFSSPy and opened issues on the repo ([see here](https://github.com/dstansby/pfsspy/issues/211)). Is this okay? If so, I will assign myself as editor.
