### 2020-10-05 #2719 @arfon
> @Kevin-Mattheus-Moerman @arfon I am happy to edit! I think I can be an objective editor, but I do want to mention that I've used PFSSPy and opened issues on the repo (see here). Is this okay? If so, I will assign myself as editor.

I think this is OK for you to go ahead and edit. Thanks for declaring this :-)
