### 2020-10-05 #2719 @mbobra
Thanks for your submission, @dstansby!

I'll work on finding reviewers. If you have any suggestions (from the [reviewer spreadsheet](https://bit.ly/joss-reviewers) or otherwise), please let me know. You can suggest them without a mention, e.g. mbobra instead of @mbobra.
