### 2020-10-05 #2719 @dstansby
Not sure what the best way to filter through the reviewer spreadsheet is, but from the bottom up the first handful of people with both python and astrophysics mentioned are: sibirrer, cxwx, JBorrow, cavestruz, sambit-giri, ashkelly
