### 2020-10-08 #2732 @whedon
[REVIEW]: pfsspy: A Python package for potential field source surface modelling
**Submitting author:** @dstansby (<a href="http://orcid.org/0000-0002-1365-1908">David Stansby</a>)
**Repository:** <a href="https://github.com/dstansby/pfsspy" target ="_blank">https://github.com/dstansby/pfsspy</a>
**Version:** v0.6.3
