### 2020-10-08 #2732 @mbobra
@openjournals/joss-eics Why isn't the pdf compiling properly from branch joss-paper? (It worked correctly in the pre-review).
