### 2020-10-08 #2732 @mbobra
👋 @mancellin @sibirrer Thank you so much for agreeing to review this submission! Whedon generated your reviewer checklists above. If you'd like more information on any of these items, please reference the [review critera](https://joss.readthedocs.io/en/latest/review_criteria.html). Also, we encourage open communication between the reviewers, submitting author, and editor. So please feel free to ask questions here on this review thread. I'm happy to help!

Also, I am not sure why Whedon is not generating the paper correctly. I'll fix this, but you can use [this pdf](https://github.com/openjournals/joss-papers/blob/joss.02719/joss.02719/10.21105.joss.02719.pdf) in the meantime.
