### 2020-10-09 e53221b
Add an error test
