### 2020-10-09 dstansby/pfsspy#239 @sibirrer
Code of conduct link from README does not work
### Description

The link to the code of conduct from the readme and from the description of this issue is not working.

This issue is essential as part of my review for JOSS.
