### 2020-10-09 dstansby/pfsspy#240 @sibirrer
Contributing guidelines lacking
### Description

There are no guidelines for contributions to pfsspy. You may want to have a short statement and can refer for a 'how-to' to sunpy if you want or use available templates from astropy that outline contributions within GitHub.

This is an essential issue for my review of the JOSS submission.
