### 2020-10-09 dstansby/pfsspy#241 @sibirrer
Move classes from __init__.py to separate files to improve readability of the code
### Description

Currently a lot of executable lines of code are in the __init__.py file, including class definitions that are highly relevant for the overall functionality of pfsspy. To improve the readability of the code, I'd suggest to keep the __init__.py file to an absolute minimum and outsource all classes and definitions to separate files with a naming convention related to the class definition and tasks being executed.

This is an essential issue as part of my review for JOSS.
