### 2020-10-09 #2732 @sibirrer
@dstansby very nice software package. I very much like the presentation of the documentation and the user examples. I don't see from my side any roadblocks from publishing pfsspy in JOSS after minor revisions indicated in my opened issues.

I opened three issues:
https://github.com/dstansby/pfsspy/issues/239
https://github.com/dstansby/pfsspy/issues/240
https://github.com/dstansby/pfsspy/issues/241
one concerning the documentation, one about the contribution guidelines and one about code formatting.
Let me know if you have questions about these issues and happy to engage in a discussion.

The installation works and the few examples I tried worked out.

There is one reference in the paper pdf that does not work. And I found a type ("Given an 2D map...")
I am not an expert in solar physics and therefore do not want to judge the scientific applicability. Given that this code has been used in several publications with noticeable use, this is prove enough.

Testing is at a good stage. There are some definitions not being tested and it might be not too much work to add suitable test functions for those as well.

In the documentation you state (wisely): "pfsspy is a very new package, so elements of the API are liable to change with the first few releases." Can you be more specific of which features are susceptible to changes or do you consider the API being stable now? Or what kind of design and feature changes the user might expect in the near future? I very much like the changelog description with the changes that are not backwards compatible in the latest version.
