### 2020-10-10 475bdaf
Move input to separate file
