### 2020-10-10 dstansby/pfsspy#241 @dstansby
Doing this was long overdue! I have moved the code in `__init__.py` out to three separate files, in the three pull requests referenced above.
