### 2020-10-10 b488e6d
Add code of conduct text
