### 2020-10-10 dstansby/pfsspy#239 @dstansby
This should have been fixed by https://github.com/dstansby/pfsspy/pull/246, where I added a copy of the SunPy code of conduct to this repository to prevent broken link issues.

When I address https://github.com/dstansby/pfsspy/issues/240 I'll also add a code of conduct link in the documentation (currently it is only linked from the README).
