### 2020-10-11 5d985b2
Add contributing guidelines
