+++ doc/source/contributing.rst
@@ -1,0 +1,36 @@
+Contributing to pfsspy
+======================
+
+Reporting Issues
+^^^^^^^^^^^^^^^^
+
+If you use pfsspy and stumble upon a problem, the best way to report it is by
+opening an `issue`_ on the GitHub issue tracker.
+This way we can help you work around the problem and hopefully fix the problem!
+
+When reporting an issue, please try to provide a short description of the issue
+with a small code sample, this way we can attempt to reproduce the error.
+Also provide any error output generated when you encountered the issue, we can
+use this information to debug the issue.
+
+Requesting new features
+^^^^^^^^^^^^^^^^^^^^^^^
+
+If there is functionality that is not currently available in pfsspy you can
+make a feature request. Please add as much information as possible regarding
+the new feature you would like to see.
+
+.. _issue: https://github.com/dstansby/pfsspy/issues
+
+Contributing code
+^^^^^^^^^^^^^^^^^
+All code contributions to pfsspy are welcome! If you would like to submit a
+contribution, please open a pull request at the pfsspy repository. It will then
+be reviewed, and if it is a useful addition merged into the main code branch.
+
+If you are new to open source development, the instructions in these links might
+be useful for getting started:
+
+* `Astropy Dev Workflow <https://docs.astropy.org/en/latest/development/workflow/development_workflow.html>`_
+* `Astropy Dev environment <https://docs.astropy.org/en/latest/development/workflow/get_devel_version.html#get-devel>`_
+* `Astropy Pull Request Example <https://docs.astropy.org/en/latest/development/workflow/git_edit_workflow_examples.html#astropy-fix-example>`_
