### 2020-10-11 9d603fe
Add link to changelog in API note
