### 2020-10-11 dstansby/pfsspy#240 @dstansby
I've added a set of guidelines in https://github.com/dstansby/pfsspy/pull/247
