### 2020-10-11 #2732 @dstansby
Thanks for taking the time to review it, and thanks for the helpful review!

> I opened three issues:
> [dstansby/pfsspy#239](https://github.com/dstansby/pfsspy/issues/239)
> [dstansby/pfsspy#240](https://github.com/dstansby/pfsspy/issues/240)
> [dstansby/pfsspy#241](https://github.com/dstansby/pfsspy/issues/241)
> one concerning the documentation, one about the contribution guidelines and one about code formatting.
> Let me know if you have questions about these issues and happy to engage in a discussion.

I think I have addressed each of these issues; see the comments I have left on the individual issues. I'll leave you to close or comment on them as you see fit. Am happy to do a minor release of pfsspy with these changes as a condition of acceptance, but I will wait for reviewer two to comment in case there's some more changes to be made.

> There is one reference in the paper pdf that does not work. And I found a type ("Given an 2D map...")

Thanks, the typo and reference has been fixed in the latest version of the paper source.

> In the documentation you state (wisely): "pfsspy is a very new package, so elements of the API are liable to change with the first few releases." Can you be more specific of which features are susceptible to changes or do you consider the API being stable now? Or what kind of design and feature changes the user might expect in the near future? I very much like the changelog description with the changes that are not backwards compatible in the latest version.

Technically I'm reserving the right for everything to be subject to change; in practice now `pfsspy` is maturing only minor changes are now made. I intend for the current version (0.6) to be the last unstable version, and for the next major release (1.0) to be the first release where no API is subsequently broken.

I think that's all - any more queries or questions let me know!
