### 2020-10-13 dstansby/pfsspy#250 @mancellin
Error in tracer_performance.py
I get the following error when trying to run `examples/pfsspy_info/tracer_performance.py`
```
Traceback (most recent call last):
  File "tracer_performance.py", line 34, in <module>
    br = sunpy.map.Map(br, pfsspy.carr_cea_wcs_header('2010-01-01', br.shape))
AttributeError: module 'pfsspy' has no attribute 'carr_cea_wcs_header'
```

I tried to replace `pfsspy` by `pfsspy.utils`, but then hit another error, so I'm guessing the function has changed somehow when it moved in the `utils` module.
