### 2020-10-13 dstansby/pfsspy#252 @mancellin
Suggestions for documentation and UI
- [x] You could mention the optional packages (`streamtracer`, `numba`) in the installation section. (For instance, for users with limited internet connection wanting to install the full package at once.)  A link to `streamtracer` homepage would also be nice in the 'Improving performance ' section.
- [ ] You could add a few paragraphs at the beginning of the documentation to explain the general structure of the code and the terminology. From the example, I am guessing from the examples that `pfsspy.pfss` is the main function solving the PDE problem and that "tracers" are post-processing tools to reconstruct the values of the magnetic field from the abstract output of `pfss`. Maybe this could be explained more explicitly. By the way, wouldn't `solve` or `solve_pfss` be a more explicit name for `pfss`?
- [ ] Adding a string representation `__str__` to `Input` and `Output` might help new users to understand what they store.

These suggestions are not mandatory for a positive JOSS review. Do not feel any obligation to implement them.
