### 2020-10-20 dstansby/pfsspy#250 @mancellin
The PR #251 fixes the issue. Thank you!
