### 2020-10-20 dstansby/pfsspy#254 @mancellin
Gong data changed?
Hi,

I'm getting the following error with the test suite:
```
ValueError: ['https://gong2.nso.edu/oQR/zqs/202009/mrzqs200901/mrzqs200901t1304c2234_022.fits.gz'] has already been downloaded, but no file matching the hash 9272fcb09131a7b0b51c088f61c75a8923b76f32309dd1c5596b484469dfbb79 can be found.
```
It was working fine a few days ago, so I'm guessing the content of the dataset has changed.
I'm not familiar with this data archive. Is it expected to happen?
