### 2020-10-20 #2732 @mancellin
This is a very good work. The package is easy to install, the documentation is fine, and the code examples work as expected.

I agree with @sibirrer's comments. I also reported two more minor issues (dstansby/pfsspy#250, dstansby/pfsspy#254) and made some optional suggestions (dstansby/pfsspy#252, dstansby/pfsspy#253) in the issue tracker of the package.

The paper can be recommended for publication as soon as a fixed version of the package is released.
