### 2020-10-22 76feb0e
Add performance deps install option
