+++ setup.cfg
@@ -57,6 +57,9 @@ test =
   pytest-cov
   streamtracer
   reproject
+performance =
+  numba
+  streamtracer
 
 # See the docstring in versioneer.py for instructions. Note that you must
 # re-run 'versioneer.py setup' after changing this section, and commit the
