### 2020-10-22 dstansby/pfsspy#252 @dstansby
Thanks for the suggestions - I have edited the issue to give them tick boxes so I can keep track of what is done. I think I will fix the first one in time for the JOSS review, but will leave the other two open as they will require a bit more thought.
