### 2020-10-22 dstansby/pfsspy#254 @dstansby
The archive shouldn't change often, but it looks like they have done a mass re-write of all the files recently, possible to fix the metadata in some global way.
