### 2020-10-22 34bff9a
Fix GONG file hash
