### 2020-10-22 #2732 @dstansby
Thanks both for your super helpful reviews. I have just released a new minor version with the recommended fixes: https://pypi.org/project/pfsspy/0.6.2/. Let me know if there is anything else I need to do, or if I have missed anything.
