### 2020-10-22 #2732 @dstansby
Thanks! Version 0.6.3 includes the paper source: https://github.com/dstansby/pfsspy/releases/tag/0.6.3 / https://pypi.org/project/pfsspy/0.6.3/
